+++ src/clj/skylobby/fx/battles_buttons.clj
@@ -80,11 +80,13 @@
            :value (fx/sub-val context :battles-layout)
            :items battles-layouts
            :on-value-changed {:event/type :spring-lobby/assoc
-                              :key :battles-layout}}
-          {:fx/type :button
-           :text "Close Battle Preview"
-           :on-action {:event/type :spring-lobby/dissoc-in
-                       :path [:by-server server-key :selected-battle]}}]
+                              :key :battles-layout}}]
+         (when (fx/sub-val context get-in [:by-server server-key :selected-battle])
+           [
+            {:fx/type :button
+             :text "Close Battle Preview"
+             :on-action {:event/type :spring-lobby/dissoc-in
+                         :path [:by-server server-key :selected-battle]}}])
          #_
          [{:fx/type :label
            :text " Resources: "}